requirements.txt
brotli==1.2.0
cffi==2.0.0
cssselect2==0.8.0
fonttools==4.61.1
pillow==12.1.0
pycparser==3.0
pydyf==0.12.1
pyphen==0.17.2
tinycss2==1.5.1
tinyhtml5==2.0.0
weasyprint==68.0
webencodings==0.5.1
zopfli==0.4.0
pikepdf==10.3.0
boto3==1.34.14
django-storages==1.14.2
requests>=2.31.0
